refactor: split god-module into core/, mixins/, outputs/, cli/ subpackages
- Extract MicroBenchBase, MicroBench, contexts, encoding, monitoring into microbench/core/
- Extract all MB* mixins into microbench/mixins/ (base, call, python, system, vcs, profiling, gpu)
- Extract Output classes into microbench/outputs/ (base, file, http, redis, utils)
- Extract CLI into microbench/cli/ (main, parser, registry, runner)
- Add microbench/version.py for __version__
- Replace __init__.py (863 lines) with thin public-API facade
- Replace __main__.py (742 lines) with thin shim to cli/main.py
- Add DeprecationWarning shims for _mixins, _output, _encoding private modules
- Public API unchanged: all root imports, CLI entrypoint, record schema stable
