+++ microbench/cli/__init__.py
@@ -1,0 +1,1 @@
+# microbench.cli — command-line interface implementation.
