+++ microbench/core/__init__.py
@@ -1,0 +1,2 @@
+# microbench.core — internal implementation package.
+# Public API is re-exported from microbench.__init__.
