+++ microbench/outputs/__init__.py
@@ -1,0 +1,2 @@
+# microbench.outputs — output sink classes.
+# Public API is re-exported from microbench.__init__.
